Added Kafka Streams wrapper for the EWMA outlier detector.
This is still work in progress demo code, but it's close enough that I
wanted to get it out there for feedback. It includes a Kafka Streams
consumer app for reading MetricPoints off a topic called "metrics" and
for writing the ones that are anomalies back to a topic called
"anomalies", but suitably tagged with outlierLevel="WEAK" or
outlierLevel="STRONG" as the case may be.

I also included a white noise generator in the form of a Kafka Streams
producer for MetricPoints. This writes a white noise series to the metrics
topic so the EWMA outlier detector can evaluate them.

Here are some known issues:

- The consumer app has only a single EWMA detector, which is totally
  wrong. Each series will have its own detector and so I'll need a map to
  store these on a per-series basis. But to make that work we will need
  unique names, and currently Haystack is using non-unique metric names
  like "latency" or "duration". So we need to have a conversation around
  whether we can make those names unique (possibly relegating "latency"
  and "duration" to a tag), or whether the unique name should itself be a
  tag.
- Config for both the producer and consumer apps needs to be externalized.
- Topic names should be configurable rather than hardcoded.

I'm sure there are other things I've neglected since I'm a Kafka noob. I
_don't_ think that we need KTable storage in this case or in most other
cases as models grow stale, and the EWMA model in particular quickly
reestablishes itself from ground zero anyway.