Refactor/metric source (#31)
* Refactored MetricSource and related classes.

This is in preparation for CsvMetricSource. There are certain features I
want to support in both WriteNoiseMetricSource and CsvMetricSource, and I
want to implement them in one place instead of many. These include:

- Multiple subscribers for a metric source, including the corresponding
  management methods (add/remove/publish).
- A timer for higher-fidelity simulation. (Clients can set the period to 0
  if desired.)

I updated the MetricSource/MetricSubscriber interfaces to these ends, and
created an AbstractMetricSource to handle the actual implementation of
what's common. I also backported the WhiteNoiseMetricSource to use the new
refactoring. So now we can create metric sources without worrying about
subscription management or timers. All we have to do is generate the next
instant and value.

* Corrected class name