Added a CSV metric source, and created a pipeline mini-framework.
For model development and testing, it's useful to be able to connect
outlier detectors to metric sources, and to do this without having to
fire up an underlying messaging infrastructure. I created a basic source/
filter/sink framework in the tools module, which allows us to do exactly
this. I used wrappers for outlier detectors to avoid polluting their
API with what's essentially Java event semantics.

There are a couple of sample pipelines in the
com.expedia.adaptivealerting.samples package.