Added band to the time series chart.
This necessitated a change that I suspected we needed anyway, which was to
have the OutlierDetector.classify() method return a MetricPoint instead of
returning an OutlierLevel. I'm returning a MetricPoint since (1) we pass a
MetricPoint to the anomaly detection component and (2) we need some way to
communicate the thresholds so they can be visualized or otherwise used.