Added Mpoint, Metric and MetricFrame, and updated samples accordingly.
This set of changes allows us to set a sub-second timer period for the
sample pipelines driven by CSV files.

Explanation: Currently, the finest resolution supported by the Scala
MetricPoint implementation is 1s. Since we are using the same MetricPoint
implementation to drive the pipeline charts, the pipeline charts have the
same finest resolution.

Originally we were using Instant.now() to set the MetricPoint timestamp.
Given the above, we had to set the period to at least 1s to avoid
multiple metric points being assigned to the same second. This made the
demos pretty slow.

With the changes, we can run the CSV file demos faster. We specify the
timestamps in the CSV file and use those to drive the charts. Then the
timer period can be whatever we want.

I used MetricFrame class to store the timestamps and data values. To
make this stream, I added the Mpoint class since the MetricPoint class is
going away at some point. The Metric class is just a placeholder til we
fill in the Metrics 2.0 tags.