Align Mpoint more closely with Haystack
Some internal integrations require a more incremental approach to
evolving the Haystack MetricPoint representation. Therefore I've updated
the AA Mpoint representation here to use epochTimeInSeconds and a Float
value, instead of using an Instant and a Double value.

I am still interested in using Instant and Double here, but we can have
that conversation after the basic integrations are working.