Created main() for KafkaAnomalyDetectorManager
The KADM needs to be able to create detectors on demand. To do this we use
the factory pattern. When the KADM starts up, it needs to know the mapping
from detector names to factory classnames. Hence the number of factory-
related classes in this commit.

In the future we want to get rid of the detector-specific KStreams apps
and just use the KADM to host detectors. But for now the system works with
detector-specific apps and I didn't want to break that. So that's all left
alone. As part of the KADM implementation, though, I did externalize the
names of the inbound and outbound topics to config files, and I did this
using the configuration property names inbound-topic and outbound-topic.
To avoid having to handle both "topic" (the old name for the inbound
topic) and "inbound-topic", I just migrated everything over to using the
"inbound-topic" name, and switch the property names over to constants to
establish better control. I've updated the Haystack Terraform to define
both "topic" and "inbound-topic" values for now. Once this AA update is
deployed, we can remove the "topic" definitions from the Haystack
Terraform.