Temporarily migrated Aquila model into Adaptive Alerting. (#100)
Eventually we want to boot this out of AA and have it in its own separate
repo. Right now it's easier to have Aquila in AA because it allows us to
avoid having to figure out how to link the Aquila JAR into the AA runtime.
It shouldn't be hard to do (some combination of JVM classpath and
Terraform config) but we can refactor once we get the flow working.