Created a new DataService abstraction, with file- and S3-based impls.
The system already had file- and S3-based access, but this access was low-
level access that required the client to handle the data lookup, query and
assembly concerns. So for example Aquila originally required logic to do
the following:

- Map a metric to its location and file format
- Decompose a date range into a list of files to be loaded
- Query the data store for the files
- Assemble the results into a single data frame

Previously, other models would need to do the same thing. For example, the
AWS SageMaker Random Cut Forest model would have to do the same.

With this commit I created a DataService to handle these concerns.
Currently the DataService is a simple JavaBean, so it's accessible only
from Java, but shortly I'll put a web service around it so we can use it
to provide data to arbitrary web clients.