Implemented CUSUM factory, and framed out the other factories.
This PR is large but most of this is just flattening the anomaly detector
packages.

The code of central interest in this PR is

  com.expedia.adaptivealerting.anomdetect.cusum

CusumModel allows us to read model files. This includes the params as well
as certain other fields (model UUID and type, which we have in the model
files to make them easier to track). The CusumFactory reads the CusumModel
from S3. (Eventually this will be abstracted away.)

TODO Rethink the factory config so that all of the factories can share the
same AWS region and S3 bucket. We don't want to define these on a per-
factory basis.