ADManager config bugfix
The ADManager config expected to see region and bucket params that don't
actually exist. Rather than adding these, I hardcoded the region and
bucket, because we plan to switch over to DB-based model storage very
shortly (in two weeks or so). When we do this, the region and bucket
params go away.