Added basic detector source in preparation for Haystack metrics. (#311)
This commit adds a basic detector source that handles both detector UUID
generation and UUID-based detector lookup. We generate UUIDs dynamically
and consistently for a given MetricDefinition. Once we have a detector
UUID, we use that to create/retrieve an in-memory EWMA detector.

The AD Mapper will use UUID generation, and the AD Manager will use UUID-
based detector lookup. This will follow in a subsequent commit.

This allows us to map EWMA detectors to any incoming metric without any
user intervention, and without going to the model service.