DRY up detector registry
Previously the detector registry was defined in multiple config files.
It's the same registry across all apps, so there's no need for that.
Moreover, there's not really any need for this to be config files, since
the detector registry is wholly dependent on the built-in detectors
available. So I just implemented a DetectorLookup class. When people
create new detectors they need to create an entry in this class.