Moved Notifier app to Kafka module.
Previously the notifier module included Kafka code, which we don't want
since we are trying to keep the core domain logic isolated from the
messaging technology choice (e.g. Kafka, Kinesis, RabbitMQ, etc.).

Instead of extricating the Kafka code from notifier, I think it will be
easier to just move the whole thing over to Kafka, and then pull the
domain logic out of the Kafka app. That way we can fix the macro situation
and then refactor into the notifier module over time. (In this PR I've
simply removed the notifier module for the time being.)

Also there was a Zookeeper conflict. The Kafka module used one version and
the Notifier unit tests used another--incompatible--version. This
incompability was preventing the unit tests from running properly in
IntelliJ. After moving the Notifier into the Kafka module, it also broke
the Kafka module unit tests. I added to the Kafka module a test dependency
to the newer version of Zookeeper, just to fix the unit tests.