Removed data connector and data service modules.
These are not part of the AA runtime. They are more for ML-based trainers.
I would prefer to keep AA tightly scoped to the detector runtime for now.
Whatever support we provide for developing trainers is probably a separate
SDK repo. That is a future discussion.

I've moved the data connector/service to the Aquila project, which is
currently the only project that uses them.