Refactoring to simplify detector initialization
The original detector initialization process was somewhat byzantine, with
initialization sometimes happening in constructors and sometimes in the
init() method. This led to a bug with Holt-Winters because we were
initializing with invalid params, and then calling params.validate() and
failing. (This commit fixes that issue.)

The AbstractAnomalyDetector base class is simpler now, and the subclasses
based on that are simpler too. They don't have to implement so many
abstract methods.

Also we were storing weak/strong sigmas in both the detector directly and
in the params object. The detector fields didn't do anything, and this was
leading to confusion about which ones do something and why the values are
different. I removed the detector fields since weak/strong sigmas are in
fact detector params.