Replaced Spring HATEOAS with custom code
We were having some trouble with the Spring HATEOAS-based Jackson module,
as the serializer doesn't have a noarg constructor, and the unit tests
were failing when trying to create a serializer. So this commit replaces
that with custom resource objects, which is nice anyway because it allows
us to avoid pulling in a bunch of Spring dependencies just for a few
resources.