Added anomaly timestamp to records produced by a2m (#428)
Originally I was using a null timestamp. I'm not sure what that translates
to, but it wouldn't be the timestamp on the underlying MetricData. So in
this commit I grab that timestamp and set it on the ProducerRecord
explicitly.

The hope is that this will align the displayed anomaly series with the
associated metric series.