Added support for tail anomalies.
The support in question is latent, meaning that the detectors themselves
support it now (at least EWMA, PEWMA and HoltWinters, in addition to the
constant threshold and CUSUM detectors, which already supported it). But
the UI and database schema need to be updated as well. I will follow up
with that in a subsequent PR.