Refactored EwmaDetector into ForecastingDetector
This commit moves us toward being able to mix and match point forecasting
algos with interval forecasting algos. Here we replace the EwmaDetector
(which coupled the EWMA point forecast algo with the exponential version
of the Welford stdev algo) with a ForecastingDetector, which takes a point
forecasting algo, an interval forecasting algo and a tail type.