Moved individuals out of legacy, since it's not a forecasting detector
Originally I thought that the individuals (Shewhart individuals control
chart) detector is a forecasting detector, but that's not exactly right.
It does generate internal forecasts, but it has its own built-in approach
to evaluating threshold breaches, and so from that perspective it's not a
forecasting algo per se; it's an anomaly detection algo.

I also added comments per Matt's requests:
https://github.com/ExpediaDotCom/adaptive-alerting/pull/466