Added detector aggregator abstraction and two implementations
This is intended to give us a way to aggregate persistent anomalies into a
transformed signal.

Specifically the m-of-n aggregator allows us to amplify a persistent
series of weak anomalies (just recent--doesn't have to be consecutive)
into a strong anomaly. This is useful in cases where we don't want to
trigger an action (e.g. automated alert or autoremediation) off a single
weak anomaly, but if the weak anomalies keep coming in then we want to
act.