Removed old KafkaAnomalyDetectorManager and related tests/config (#535)
Also:
- Updated Docker config to use KafkaDetectorManager (the new class)
- Updated Terraform config to do the same

Note that I left the app name as "ad-manager" as I didn't want to have to
go through all the places in the app and deployment scripts replacing all
of this until we're sure it's working properly. At some point we can look
at renaming the apps to detector-manager and detector-mapper. Not urgent.