Created a naive forecaster per issue #559
See https://otexts.com/fpp2/simple-methods.html#simple-methods

I have not yet created the factory for this detector, so we can't quite
use it yet. This will require a generic ForecastingDetectorFactory, which
is a little more involved since we want to have per-algo factories for
the individual forecasting algos too. Will take care of this later when I
do the seasonal naive forecaster, since that one we want to actually try
using.