Simplified the code for building detectors from documents
Previously we had a DetectorRegistry that returned an algo-specific
DetectorFactory for a given document, and clients would use that to build
a detector. Inside the DetectorRegistry there was a DetectorFactoryBuilder
interface for building detector factories. It was all more complicated
than necessary.

Instead I renamed DetectorRegistry to DetectorFactory, and treat all the
algo-specific components as providers. I got rid of DetectorFactoryBuilder
since it wasn't doing anything useful.